refactor: introduce directory/package structure for pkg source
