+++ pkg/product/test/metadata_test_template.yaml
